refactor(ui/snap-account): use PascalCase naming for snap account message components
