+++ ui/pages/confirmations/components/snap-account-error-message/SnapAccountErrorMessage.tsx
