+++ ui/pages/confirmations/components/snap-account-success-message/SnapAccountSuccessMessage.tsx
