+++ ui/pages/confirmations/components/snap-account-success-message/index.tsx
@@ -1,1 +1,1 @@
-export { default as SnapAccountSuccessMessage } from './snap-account-success-message';
+export { default as SnapAccountSuccessMessage } from './SnapAccountSuccessMessage';
